requirements.txt
aniso8601==9.0.1
astroid==3.3.3
blinker==1.8.2
click==8.1.7
dill==0.3.8
flake8==7.1.1
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==24.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
pylint==3.3.0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.35
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
Werkzeug==3.0.4
